Added assets
I've added in the actual game map, that the player-character is gonna be moving around.

I've then added a json file, which defines collision boxes, so that I can use that for code later.

Also added the same player-character, which I used for my other game, since these games are both gonna be used for my portfólio.
Also added a broken heart, in case I want to show the character dying or something, Idk.

I added a png for the foreground objects, so that the character will be able to walk underneath of those.

*What's next?

I might add a js file, which defines the holes, since I want to use those to make different cool things to show off.